# backend/services/calcsvc/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Calculator Service
  description: HTTP service for adding numbers, a goa teaser
  version: ""
host: localhost:8082
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /calcsvc/v1/add/{a}/{b}:
    get:
      tags:
      - calcsvc
      summary: add calcsvc
      operationId: calcsvc#add
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: integer
      - name: b
        in: path
        description: Right operand
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /calcsvc/v1/multiply:
    post:
      tags:
      - calcsvc
      summary: multiply calcsvc
      operationId: calcsvc#multiply
      parameters:
      - name: MultiplyRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/CalcsvcMultiplyRequestBody'
          required:
          - a
          - b
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /swagger.json:
    get:
      summary: Download ../../gen/http/openapi.json
      operationId: openapi#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
definitions:
  CalcsvcMultiplyRequestBody:
    title: CalcsvcMultiplyRequestBody
    type: object
    properties:
      a:
        type: integer
        description: Multiplicand
        example: 4288748512599820841
        format: int64
      b:
        type: integer
        description: Multiplier
        example: 4212629202012168060
        format: int64
    example:
      a: 1698882017578366363
      b: 6747375795581831989
    required:
    - a
    - b